add: fix for data object and moved forkHandle implementation into forkHandle.h
