+++ include/ForkHandle.h
@@ -19,13 +19,6 @@
  * @version 2.02, 8/14/2014
  */
 
-/*
- *  fork.h
- *  
- *
- *  Created by Jiaxi Jin on 6/18/10.
- *
- */
 
 #ifndef FORKHANDLE_H
 #define FORKHANDLE_H
@@ -35,7 +28,60 @@
 #include <signal.h>
 #include "GenSync.h"
 
-int forkHandle(GenSync* client1, GenSync* client2, long &bytes, double &communTime, double &totalTime);
 
 
+/**
+ * Report structure for a forkHandle run
+  */
+struct forkHandleReport {
+    forkHandleReport(): bytes(-1), CPUtime(-1), totalTime(-1), success(false) {}
+    long bytes;       // the number of bytes communicated
+    double CPUtime;   // the amount of CPU time used
+    double totalTime; // total time used
+    bool success;     // true iff the sync completed successfully
+};
+
+/**
+ * Runs client (child process) and server (parent process), returning statistics for server.
+ * server is modified to reflect reconciliation, whereas client is not.
+ * @param server The GenSync object that plays the role of server in the sync.
+ * @param client The GenSync object that plays the role of client in the sync.
+ * @return Synchronization statistics as reported by the server.
+ */
+inline forkHandleReport forkHandle(GenSync& server, GenSync client) {
+    int err = 1;
+    int chld_state;
+    int my_opt = 0;
+    forkHandleReport result;
+    clock_t start = clock();
+    try {
+        pid_t pID = fork();
+        int method_num = 0;
+        if (pID == 0) {
+            signal(SIGCHLD, SIG_IGN);
+            Logger::gLog(Logger::COMM,"created a server process");
+            client.listenSync(method_num);
+            exit(0);
+        } else if (pID < 0) {
+            //handle_error("error to fork a child process");
+            cout << "throw out err = " << err << endl;
+            throw err;
+        } else {
+            Logger::gLog(Logger::COMM,"created a client process");
+            server.startSync(method_num);
+            result.totalTime = (double) (clock() - start) / CLOCKS_PER_SEC;
+            result.CPUtime = server.getSyncTime(method_num); /// assuming method_num'th communicator corresponds to method_num'th syncagent
+            result.bytes = server.getXmitBytes(method_num) + server.getRecvBytes(method_num);
+            waitpid(pID, &chld_state, my_opt);
+            result.success=true;
+        }
+
+    } catch (int& err) {
+        sleep(1); // why?
+        cout << "handle_error caught" << endl;
+        result.success=false;
+    }
+    
+    return result;
+}
 #endif
